+++ common/routes/Video/components/BlockItems9.js
@@ -13,7 +13,7 @@ class BlockItems9 extends PureComponent {
   select (index) {
     if (this.props.selectVideo !== index) {
       this.props.selectVideo(index);
-      window.document.body.scrollTop = 0;
+      window.document.documentElement.scrollTop = 0;
     }
   }